styles(sidebar): migrate to styled-components
